# docs/microflow/contracts/openapi-draft.yaml
openapi: 3.0.3
info:
  title: Atlas Microflow API (draft)
  version: "0.21.0"
  description: |
    与 `mendix-studio-core` 微流契约第 21 轮对齐。Authoring 模型见 `docs/lowcode-mendix-mvp/schema.md` 与
    `@atlas/microflow` 类型；不存储 FlowGram JSON。
servers:
  - url: /
paths:
...
  /api/microflows/{id}/rename:
    post:
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                displayName: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeMicroflowResource" }
...
  /api/microflows/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      summary: Get resource
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeMicroflowResource" }
    patch:
      summary: Patch resource
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                patch: { $ref: "#/components/schemas/MicroflowResourcePatch" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeMicroflowResource" }
    delete:
      summary: Delete resource
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvelopeId" }
...
components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    MicroflowApiResponse:
      type: object
      required: [success, timestamp]
      properties:
        success: { type: boolean }
        data: {}
        error: { $ref: "#/components/schemas/MicroflowApiError" }
        traceId: { type: string }
        timestamp: { type: string, format: date-time }
    MicroflowApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - MICROFLOW_NOT_FOUND
            - MICROFLOW_NAME_DUPLICATED
            - MICROFLOW_SCHEMA_INVALID
            - MICROFLOW_VALIDATION_FAILED
            - MICROFLOW_VERSION_CONFLICT
            - MICROFLOW_PUBLISH_BLOCKED
            - MICROFLOW_REFERENCE_BLOCKED
            - MICROFLOW_PERMISSION_DENIED
            - MICROFLOW_ARCHIVED
            - MICROFLOW_RUN_FAILED
            - MICROFLOW_RUN_CANCELLED
            - MICROFLOW_METADATA_NOT_FOUND
            - MICROFLOW_STORAGE_ERROR
            - MICROFLOW_UNKNOWN_ERROR
        message: { type: string }
        details: { type: string }
        validationIssues: { type: array, items: { type: object, additionalProperties: true } }
        retryable: { type: boolean }
...
    MicroflowResource:
      type: object
      description: 见 TS MicroflowResource
      additionalProperties: true
    MicroflowResourcePatch:
      type: object
      additionalProperties: true
...
    EnvelopeMicroflowResource:
      allOf:
        - $ref: "#/components/schemas/MicroflowApiResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/MicroflowResource" }
    EnvelopeId:
      allOf:
        - $ref: "#/components/schemas/MicroflowApiResponse"
        - type: object
          properties:
            data:
              type: object
              required: [id]
              properties: { id: { type: string } }
...